Connection to firebase will display throught out the activity
