+++ app/src/main/AndroidManifest.xml
@@ -37,6 +37,7 @@
 
         <activity
             android:name=".ReplyActivity"
+            android:configChanges="orientation|screenSize|keyboardHidden"
             android:label="@string/app_name" >
         </activity>
