+++ app/src/main/java/hk/ust/cse/hunkim/questionroom/JoinActivity.java
@@ -14,6 +14,12 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.TextView;
+import android.widget.Toast;
+
+import com.firebase.client.DataSnapshot;
+import com.firebase.client.Firebase;
+import com.firebase.client.FirebaseError;
+import com.firebase.client.ValueEventListener;
 
 
 /**
@@ -25,13 +31,35 @@ public class JoinActivity extends AppCompatActivity implements ActionBar.TabList
 
     private ViewPager viewPager;
     private TabsPagerAdapter mAdapter;
+    private Firebase mFirebaseRef;
+    private ValueEventListener mConnectedListener;
     // Tab titles
     private String[] tabs = { "Join Room", "Room List", "Create Room" };
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
+        Firebase.setAndroidContext(this);
         setContentView(R.layout.activity_join);
+         mFirebaseRef = new Firebase(FIREBASE_URL);
+        mConnectedListener = mFirebaseRef.getRoot().child(".info/connected").addValueEventListener(new ValueEventListener() {
+            @Override
+            public void onDataChange(DataSnapshot dataSnapshot) {
+                if(dataSnapshot.getValue()!=null){
+                    boolean connected = (Boolean) dataSnapshot.getValue();
+                    if (connected) {
+                        Toast.makeText(JoinActivity.this, "Connected to Firebase", Toast.LENGTH_SHORT).show();
+                    } else {
+                        Toast.makeText(JoinActivity.this, "Disconnected from Firebase", Toast.LENGTH_SHORT).show();
+                    }
+                }
+            }
+
+            @Override
+            public void onCancelled(FirebaseError firebaseError) {
+                // No-op
+            }
+        });
 
         mAdapter = new TabsPagerAdapter(getSupportFragmentManager());
 
@@ -55,6 +83,12 @@ public class JoinActivity extends AppCompatActivity implements ActionBar.TabList
                             .setTabListener(this));
         }
 
+    }
+
+    @Override
+    protected void onDestroy() {
+        super.onDestroy();
+        mFirebaseRef.getRoot().child(".info/connected").removeEventListener(mConnectedListener);
     }
 
     @Override
